Provide example of keyword/type applicability.
The word "applicability" is a good name for the key concept of this
section.  The title feels rather vague otherwise.

Make clear that this is a natural consequence of keyword independence,
which is explained in the immediately previous section.  Otherwise,
having a section about this makes it sound like its own special case.

Use "multipleOf" to show why this is an essential behavior.  When you
need to use a type-specific keyword, and the negation of a keyword
specific to the same type, it only works because type is not automatically
constrained.

I used "multipleOf" because (unlike "maxLength"), you cannot eliminate
the use of "not" by changing keywords (to "minLength").  There is no
other way in JSON schema to express not being a multiple of something.