Incorporate review feedback: better example
This example simplifies the concepts by getting rid of the
confusing variable-length URIs and just acknowledging that the
parent URI for the root node does not make any sense.  This
sort of situation is addressed elsewhere in the spec, so add
an xref to it.

The reworked example uses two links: one to illustrate standard
resolution (not appearing in "hrefPointers") alongside of
an "hrefPointers" adjustment using an absolute pointer, and then
another illustratig both absolute and relative pointers.