$dynamicAnchor creates fragment, no base URI stuff
Instead of using a boolean (which never really fit with how
the similarly named "$anchor" worked, and meant only one
dynamic resolution was possible at a time) and messing
with base URIs and re-resolving URI-references (a topic
that creates plenty of confusion even on its own),
use a fragment name the same as "$anchor" and make "$dynamicRef"
work by just looking for the outermost resource with a matching
"$dynamicAnchor" fragment name.

This stil requires that the initial "$dynamicRef" target URI
be a valid URI, with a fragment created by "$dynamicAnchor",
in order to produce the dynamic behavior.  This ensures
that it is always clear which schemas are dynamic extension
points and which are not.